requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
python-dotenv==1.0.1
xhtml2pdf==0.2.16
Werkzeug==3.0.6
